cpu_logger/requirements.txt
kafka-python==2.0.2
psutil==5.9.4